Move the BehaviorOphysJsonApi
The commit moves the BehaviorOphysJsonApi defined in
allensdk.brain_observatory.behavior.write_nwb.__main__ to
allensdk.brain_observatory.behavior.session_apis.data_io.behavior_ophys_json_api.

This commit also fixes a bug with multiple inheritence and method
resolution order. ABC classes should always be inherited last so that
the first parent of a class will be able to fill the subclass with
defined methods (and not abstract methods).
